Move body serialization into pipeline
