+++ sdk/core/core-client/review/core-client.api.md
@@ -42,6 +42,7 @@ export interface ClientPipelineOptions extends InternalPipelineOptions {
         credential: TokenCredential;
     };
     deserializationOptions?: DeserializationPolicyOptions;
+    serializationOptions?: serializationPolicyOptions;
 }
 
 // @public (undocumented)
@@ -223,6 +224,7 @@ export type OperationRequest = PipelineRequest<OperationRequestInfo>;
 
 // @public
 export interface OperationRequestInfo {
+    operationArguments?: OperationArguments;
     operationResponseGetter?: (operationSpec: OperationSpec, response: PipelineResponse) => undefined | OperationResponseMap;
     operationSpec?: OperationSpec;
     shouldDeserialize?: boolean | ((response: PipelineResponse) => boolean);
@@ -310,6 +312,18 @@ export interface SequenceMapperType {
 }
 
 // @public
+export function serializationPolicy(options?: serializationPolicyOptions): PipelinePolicy;
+
+// @public
+export const serializationPolicyName = "serializationPolicy";
+
+// @public
+export interface serializationPolicyOptions {
+    serializerOptions?: SerializerOptions;
+    stringifyXML?: (obj: any, opts?: XmlOptions) => string;
+}
+
+// @public
 export interface Serializer {
     // (undocumented)
     deserialize(mapper: Mapper, responseBody: any, objectName: string, options?: SerializerOptions): any;
@@ -335,7 +349,7 @@ export class ServiceClient {
     constructor(options?: ServiceClientOptions);
     sendOperationRequest(operationArguments: OperationArguments, operationSpec: OperationSpec): Promise<OperationResponse>;
     sendRequest(request: PipelineRequest): Promise<PipelineResponse>;
-    }
+}
 
 // @public
 export interface ServiceClientOptions {
