+++ sdk/core/core-client/src/serviceClient.ts
@@ -22,18 +22,15 @@ import {
   FullOperationResponse,
   DictionaryMapper,
   CompositeMapper,
-  XmlOptions,
-  XML_CHARKEY,
-  XML_ATTRKEY,
-  RequiredSerializerOptions
+  XmlOptions
 } from "./interfaces";
 import { getPathStringFromParameter, isStreamOperation } from "./interfaceHelpers";
-import { MapperTypeNames } from "./serializer";
 import { getRequestUrl } from "./urlHelpers";
 import { isPrimitiveType } from "./utils";
 import { getOperationArgumentValueFromParameter } from "./operationHelpers";
 import { deserializationPolicy, DeserializationPolicyOptions } from "./deserializationPolicy";
 import { URL } from "./url";
+import { serializationPolicy, serializationPolicyOptions } from "./serializationPolicy";
 
 /**
  * Options to be provided while creating the client.
@@ -94,11 +91,6 @@ export class ServiceClient {
   private readonly _requestContentType?: string;
 
   /**
-   * Decoupled method for processing XML into a string.
-   */
-  private readonly _stringifyXML?: (obj: any, opts?: XmlOptions) => string;
-
-  /**
    * The HTTP client that will be used to send requests.
    */
   private readonly _httpsClient: HttpsClient;
@@ -115,14 +107,14 @@ export class ServiceClient {
     this._requestContentType = options.requestContentType;
     this._baseUri = options.baseUri;
     this._httpsClient = options.httpsClient || new DefaultHttpsClient();
-    this._stringifyXML = options.stringifyXML;
     const credentialScopes = getCredentialScopes(options);
     this._pipeline =
       options.pipeline ||
       createDefaultPipeline({
         credentialScopes,
         credential: options.credential,
-        parseXML: options.parseXML
+        parseXML: options.parseXML,
+        stringifyXML: options.stringifyXML
       });
   }
 
@@ -160,6 +152,7 @@ export class ServiceClient {
     request.method = operationSpec.httpMethod;
     request.additionalInfo = {};
     request.additionalInfo.operationSpec = operationSpec;
+    request.additionalInfo.operationArguments = operationArguments;
 
     const contentType = operationSpec.contentType || this._requestContentType;
     if (contentType) {
@@ -231,8 +224,6 @@ export class ServiceClient {
       }
     }
 
-    serializeRequestBody(request, operationArguments, operationSpec, this._stringifyXML);
-
     if (request.streamResponseBody === undefined) {
       request.streamResponseBody = isStreamOperation(operationSpec);
     }
@@ -252,152 +243,12 @@ export class ServiceClient {
   }
 }
 
-/**
- * @internal @ignore
- */
-export function serializeRequestBody(
-  request: OperationRequest,
-  operationArguments: OperationArguments,
-  operationSpec: OperationSpec,
-  stringifyXML: (obj: any, opts?: XmlOptions) => string = function() {
-    throw new Error("XML serialization unsupported!");
-  }
-): void {
-  const serializerOptions = operationArguments.options?.serializerOptions;
-  const updatedOptions: RequiredSerializerOptions = {
-    xml: {
-      rootName: serializerOptions?.xml.rootName ?? "",
-      includeRoot: serializerOptions?.xml.includeRoot ?? false,
-      xmlCharKey: serializerOptions?.xml.xmlCharKey ?? XML_CHARKEY
-    }
-  };
-
-  const xmlCharKey = updatedOptions.xml.xmlCharKey;
-  if (operationSpec.requestBody && operationSpec.requestBody.mapper) {
-    request.body = getOperationArgumentValueFromParameter(
-      operationArguments,
-      operationSpec.requestBody
-    );
-
-    const bodyMapper = operationSpec.requestBody.mapper;
-    const {
-      required,
-      serializedName,
-      xmlName,
-      xmlElementName,
-      xmlNamespace,
-      xmlNamespacePrefix
-    } = bodyMapper;
-    const typeName = bodyMapper.type.name;
-
-    try {
-      if (request.body || required) {
-        const requestBodyParameterPathString: string = getPathStringFromParameter(
-          operationSpec.requestBody
-        );
-        request.body = operationSpec.serializer.serialize(
-          bodyMapper,
-          request.body,
-          requestBodyParameterPathString,
-          updatedOptions
-        );
-
-        const isStream = typeName === MapperTypeNames.Stream;
-
-        if (operationSpec.isXML) {
-          const xmlnsKey = xmlNamespacePrefix ? `xmlns:${xmlNamespacePrefix}` : "xmlns";
-          const value = getXmlValueWithNamespace(
-            xmlNamespace,
-            xmlnsKey,
-            typeName,
-            request.body,
-            updatedOptions
-          );
-
-          if (typeName === MapperTypeNames.Sequence) {
-            request.body = stringifyXML(
-              prepareXMLRootList(
-                value,
-                xmlElementName || xmlName || serializedName!,
-                xmlnsKey,
-                xmlNamespace
-              ),
-              { rootName: xmlName || serializedName, xmlCharKey }
-            );
-          } else if (!isStream) {
-            request.body = stringifyXML(value, {
-              rootName: xmlName || serializedName,
-              xmlCharKey
-            });
-          }
-        } else if (
-          typeName === MapperTypeNames.String &&
-          (operationSpec.contentType?.match("text/plain") || operationSpec.mediaType === "text")
-        ) {
-          // the String serializer has validated that request body is a string
-          // so just send the string.
-          return;
-        } else if (!isStream) {
-          request.body = JSON.stringify(request.body);
-        }
-      }
-    } catch (error) {
-      throw new Error(
-        `Error "${error.message}" occurred in serializing the payload - ${JSON.stringify(
-          serializedName,
-          undefined,
-          "  "
-        )}.`
-      );
-    }
-  } else if (operationSpec.formDataParameters && operationSpec.formDataParameters.length > 0) {
-    request.formData = {};
-    for (const formDataParameter of operationSpec.formDataParameters) {
-      const formDataParameterValue = getOperationArgumentValueFromParameter(
-        operationArguments,
-        formDataParameter
-      );
-      if (formDataParameterValue !== undefined && formDataParameterValue !== null) {
-        const formDataParameterPropertyName: string =
-          formDataParameter.mapper.serializedName || getPathStringFromParameter(formDataParameter);
-        request.formData[formDataParameterPropertyName] = operationSpec.serializer.serialize(
-          formDataParameter.mapper,
-          formDataParameterValue,
-          getPathStringFromParameter(formDataParameter),
-          updatedOptions
-        );
-      }
-    }
-  }
-}
-
-/**
- * Adds an xml namespace to the xml serialized object if needed, otherwise it just returns the value itself
- */
-function getXmlValueWithNamespace(
-  xmlNamespace: string | undefined,
-  xmlnsKey: string,
-  typeName: string,
-  serializedValue: any,
-  options: RequiredSerializerOptions
-): any {
-  // Composite and Sequence schemas already got their root namespace set during serialization
-  // We just need to add xmlns to the other schema types
-  if (xmlNamespace && !["Composite", "Sequence", "Dictionary"].includes(typeName)) {
-    const result: any = {};
-    result[options.xml.xmlCharKey] = serializedValue;
-    result[XML_ATTRKEY] = { [xmlnsKey]: xmlNamespace };
-    return result;
-  }
-
-  return serializedValue;
-}
-
 function createDefaultPipeline(
   options: {
     credentialScopes?: string | string[];
     credential?: TokenCredential;
     parseXML?: (str: string, opts?: XmlOptions) => Promise<any>;
+    stringifyXML?: (obj: any, opts?: XmlOptions) => string;
   } = {}
 ): Pipeline {
   const credentialOptions =
@@ -409,6 +260,9 @@ function createDefaultPipeline(
     credentialOptions,
     deserializationOptions: {
       parseXML: options.parseXML
+    },
+    serializationOptions: {
+      stringifyXML: options.stringifyXML
     }
   });
 }
@@ -427,6 +281,10 @@ export interface ClientPipelineOptions extends InternalPipelineOptions {
    * Options to customize deserializationPolicy.
    */
   deserializationOptions?: DeserializationPolicyOptions;
+  /**
+   * Options to customize serializationPolicy.
+   */
+  serializationOptions?: serializationPolicyOptions;
 }
 
 /**
@@ -446,29 +304,12 @@ export function createClientPipeline(options: ClientPipelineOptions = {}): Pipel
     );
   }
 
+  pipeline.addPolicy(serializationPolicy(options.serializationOptions), { phase: "Serialize" });
   pipeline.addPolicy(deserializationPolicy(options.deserializationOptions), {
     phase: "Deserialize"
   });
 
   return pipeline;
-}
-
-function prepareXMLRootList(
-  obj: any,
-  elementName: string,
-  xmlNamespaceKey?: string,
-  xmlNamespace?: string
-): { [key: string]: any[] } {
-  if (!Array.isArray(obj)) {
-    obj = [obj];
-  }
-  if (!xmlNamespaceKey || !xmlNamespace) {
-    return { [elementName]: obj };
-  }
-
-  const result = { [elementName]: obj };
-  result[XML_ATTRKEY] = { [xmlNamespaceKey]: xmlNamespace };
-  return result;
 }
 
 function flattenResponse(
